Update cpp example to use vector<double>, correct CMakeLists.txt (#196)
Former-commit-id: f9674d75988f035f1e7d1740a5f890000d1ab910

+++ examples/c++/Seed.cpp
@@ -45,7 +45,7 @@ int main(){
     game->init();
 
     // Define some actions.
-    std::vector<int> actions[3];
+    std::vector<double> actions[3];
     actions[0] = {1, 0, 0};
     actions[1] = {0, 1, 0};
     actions[2] = {0, 0, 1};
